PORTAL-1871: Update ping test’s validation string (#7557)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1871

## ℹ️ Context for reviewers

We no longer say "Create a worker account" on the landing page. Instead, we say "Create an account to apply". 

We need to update the synthetic monitor so it stops pinging the Portal on-call engineer. I've updated this to search for something more stable, like Paid Family and Medical Leave