INFRA-1120 Redirect ecs_failure_monitor from Slack to Teams (#7551)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1110
Subtask: https://lwd.atlassian.net/browse/INFRA-1120

🛠 Changes
Remove references to Slack (slackbot, JSON payload)
Add Teams uri to the lambda env variables
refactor lambda handler catch
ℹ️ Context for reviewers
This PR is part of an effort to migrate alerts and other notifications from Slack to Teams (a tool approved for the Commonwealth). These notifications will be found in the mass-pfml-deploys-shared channel in Teams. Other similar migrated alerts will be found in Teams channels with the same name as their Slack counterparts. It is recommended that you pin this Teams thread

In addition, this PR is going to look a little different than its sister tasks. The ecs_failure_monitor is a lambda that's triggered by a CloudWatch Event Rule searching for ECS tasks that failed to start. The lambda used to send messages to slack via Terriyay informing the team of failure. This PR refactored the lambda a little bit so that messages are sent to Teams instead.