API-2431 date range spans benefit years (#7590)
🎫 Ticket
https://lwd.atlassian.net/browse/api-2431

🛠 Changes
Add a new search term end_date_within to the benefit years search endpoint. This takes in a date range and only returns benefit years where the end date is within that range.

ℹ️ Context for reviewers
This wasn't originally planned for, but when I started to work on API-2431 I realized we'd be making this check if an application's dates cross the end of a benefit year in a few different places. Handling this check via the API will simplify the implementation on the front end, and keep more of the business logic in the backend of the codebase.