make e2e failure messages more clear (#7640)
the Teams webhook doesn't have the ability to thread messages. (that requires a different process that will be linked in the documentation section). This resulted in E2E failure messages generated from the Portal and API actions that looked identical.This PR was made at a request to make things more clear for folks monitoring the deploys channel.