PFMLPB-3740: Collecting Overpayments - Full Demand (#7707)
In order to test collecting OP, a claim must be in a "Complete" status, meaning the claim end date is before the date the test is being run. Additionally, the claim must have payments. Using the claims in the portal account used for testing payment statuses is a good solution since those accounts have a backlog of claims that fit the testing requirements and this account only submits one type of claim.

Before using a claim, the response from api/v1/claims/<FINEOS_ID> endpoint is used to validate that the claim is approved, completed, and has payments made. After a claim has been selected it is withdrawn in an after() hook to prevent a claim from being used in further testing.