update for DT4 (#7735)
This PR updates the flow for denying a second leave request for long->DT4. From further testing the version in DT4 displays text indicating the second leave request has been denied which is not the case in lower Fineos versions. Because of this we can assert completion of the denial using text in UI (see below).