Remove the verify task in the FINEOS intake process (#7727)
Remove the verifyOccupation() from the FINEOS intake form either with .startCreateNotification() or .createNotification(). This has been disabled in the FINEOS upgrade so the employers do not need to be reverified with a period of time anymore. The link will still be enabled, but nothing should break the FINEOS intake from creating the Absence Case.