INFRA-1135 Add monitoring and pfml-aws Modules to Verbose Parakeet (#7639)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1135

🛠 Changes
added a new script that generates a page in VP
add makefile in infra
update docs-deploy action
ℹ️ Context for reviewers
Currently, if infra/monitoring or infra/pfml-aws are the only changes in the PR, VP will not pick it up. When it's not picked up in VP, the release manager needs to be manually notified of the change.
The pages in VP are created by the generate-release-docs.sh script. This script runs commands such as make release-notes and formats the output in markdown for GitHub.