INFRA-1212 Scripted Releases Pulls Unexpected Tags (#7237)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1212

🛠 Changes
#7167 added a change that would fix how tags were collected in testing scenarios:

The finalize_release task uses the git describe command to grab the most recent tag, and then strips the prerelease data from it to mark it ready for production. However, due to the lack of commits during testing, a common result is that the HEAD of a particular branch gets multiple finalized tags
This PR updates the search algo for tags and makes logging OS agnostic.