INFRA-1257 Add BI to Verbose Parakeet (#7809)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1257

🛠 Changes
added BI changes to the Infra page on Verbose Parakeet
ℹ️ Context for reviewers
The Verbose Parakeet page needs to be updated to generate info when the infra/bi directory is changed. Currently, if that directory is the only changes in the PR, VP will not pick it up. When it's not picked up in VP, the release manager needs to be manually notified of the change.

addendum: Some of the work here is similar to the work found in #7734. The proposal here is that, in order to have docs generated by generate-release-docs an application must have tags and release branches (e.g. deploy/api/[some_version]) and lastly, contain application changes (infrastructure relating to these applications can be included as well). In order for docs to be generated by generate-infra-docs the application must have most of its functionality defined in /infra and mostly impact infrastructure changes. Tags and branches are optional here.

Following the criteria listed above. BI work should be listed under the Infra page because: the work is almost completely defined in /infra, changes in this directory are infrastructure changes rather than application changes and while BI has tags, there are no release branches.