PFMLPB-3839: Update 'review' button selector for DT4 (needs to be backwards compatible) (#7821)
This PR updates the change request flow for both approval & denial for DT4 (April release) to be backwards compatible. The updated flow for DT4 includes an additional module that pops up after clicking the "Review" button. Also accounts for approving and/or denying a claim after hitting the review button. These changes are needed as in older versions of Fineos the approve/deny button is available from the leave details page.