PORTAL-1996: ensure names don't overlap in claims search test (#7832)
https://lwd.atlassian.net/browse/PORTAL-1996

🛠 Changes

    Explicitly define names used in claims search tests

ℹ️ Context for reviewers

In our search tests, we're using other_employee.first_name as a search parameter. If other_employee.first_name matches or is a subset of employee.first_name or employee.last_name, we may receive claims from a different employee, which skews the test result and can cause intermittent failures.

This explicitly defines the names so that tests are more deterministic.