[INFRA-1247...1248] Post BI and Admin Portal Deployments to Teams (#7843)
🎫 Ticket
https://lwd.atlassian.net/browse/1247
https://lwd.atlassian.net/browse/1248

🛠 Changes
updated the data value in bi-infra-deploy
added a job that posts to teams in admin-portal-deploy
ℹ️ Context for reviewers
This PR is part of an effort to migrate alerts and other notifications from Slack to Teams (a tool approved for the Commonwealth). These notifications will be found in the mass-pfml-deploys-shared channel in Teams. It is recommended that you pin this Teams thread