INFRA-1315 Automatically Create Release Notes (#7853)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1315

🛠 Changes
add create-release-notes job to the scripted releases workflow
ℹ️ Context for reviewers
The goal of the ticket is to remove the last manual step of the release process: Generating and sharing release notes.
Following up on a request from the very early days of the scripted releases project; this PR allows users to automatically generate release notes after running a release task. GitHub makes this process pretty easy with the gh binary that is preinstalled in every github actions machine. (More information in documentation section).