API-2635: Fix how employer contributions are stored (#7939)
🎫 Ticket
https://lwd.atlassian.net/browse/api-2635

🛠 Changes
Correct a typo that was causing us to store employer contributions incorrectly.

ℹ️ Context for reviewers
The impact is minimal - we’ll just calculate this value every time it’s needed, but once we fix this we’ll be able to reuse the stored values properly instead of calculating it every time. We'll clean up the existing data separately.