PFMLPB-4054: Create PR to add E2E team to CODEOWNERS file for other important folders (#7958)
https://lwd.atlassian.net/browse/PFMLPB-3808
Subtask: https://lwd.atlassian.net/browse/PFMLPB-4054

The E2E team has implemented a lot of code w/in folders in which we don't have permission to approve and merge PRs 😞

This unfortunately is causing a bottleneck in regard to the PR approval/review/merge flow. The infra team is quite busy and may not always have the bandwidth to review PRs as needed by the E2E team. It is very important we have the ability to submit and approve PRs in the event an error occurs or quick implementation is needed in regard to deployment workflows.

I'm hoping this will not be a huge ask as the E2E team does have important code w/in these folders; and not to mention upcoming implementation to help improve the deploy process 🙏